Move FundingCycle modals to FundingCycle/modals
